pyproject.toml
[project]
name = "reminder"
version = "0.1.0"
description = "Telegram bot for cron-based reminders"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "apscheduler>=3.11.1",
    "authlib>=1.6.5",
    "dotenv>=0.9.9",
    "fastapi[standard-no-fastapi-cloud-cli]>=0.121.2",
    "google-auth>=2.43.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "marimo>=0.17.8",
    "pydantic>=2.12.4",
    "pydantic-ai-slim[google]>=1.18.0",
    "pydantic-settings>=2.12.0",
    "python-telegram-bot>=22.5",
    "sqlalchemy>=2.0.0",
]

[dependency-groups]
dev = [
    "mypy>=1.0.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.5",
    "vulture>=2.14",
]

[tool.vulture]
exclude = [".venv", ".ruff_cache", "__pycache__"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.isort]
force-sort-within-sections = true
split-on-trailing-comma = true

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
disallow_untyped_decorators = false
no_implicit_optional = false
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true
show_error_codes = true

